dto구조 변경 및 grade enum 추가를 통한 서비스와 유틸 함수 분리

+++ src/main/java/com/airqualityservice/infrastructure/api/seoul/SeoulAirQualityApiDto.java
@@ -63,9 +63,9 @@ public class SeoulAirQualityApiDto {
         @JsonProperty("ARPLT_MAIN")
         private String mainMaterial;
         @JsonProperty("PM10")
-        private Float pm10;
+        private Integer pm10;
         @JsonProperty("PM25")
-        private Float pm25;
+        private Integer pm25;
         @JsonProperty("O3")
         private Float o3;
         @JsonProperty("NO2")
